requirements.txt
google-cloud-speech
pyaudio
python-dotenv
setuptools
webrtcvad
numpy
